app: move loose ts files into util folder
