+++ app/src/lib/Util/interfaces.ts
